[Refactor] 수정 삭제 요소 분리
